+++ Source/CMakeLists.txt
@@ -3,6 +3,8 @@ set(LTL_DIR ${CMAKE_CURRENT_SOURCE_DIR}/LuaTemplateLibrary)
 set(LTL_HEADERS
     ${LTL_DIR}/Internal.hpp
     ${LTL_DIR}/FuncArguments.hpp
+    ${LTL_DIR}/LuaTypes.hpp
+    ${LTL_DIR}/STDContainers.hpp
     ${LTL_DIR}/FuncUtils.hpp
     ${LTL_DIR}/LuaAux.hpp
     ${LTL_DIR}/UserData.hpp
@@ -11,7 +13,6 @@ set(LTL_HEADERS
     ${LTL_DIR}/Class.hpp
     ${LTL_DIR}/LuaFunctions.hpp
     ${LTL_DIR}/LuaState.hpp
-    ${LTL_DIR}/LuaTypes.hpp
     ${LTL_DIR}/RefObject.hpp
     ${LTL_DIR}/StackObject.hpp
     ${LTL_DIR}/Exception.hpp
