+++ Source/LuaTemplateLibrary/LTL.hpp
@@ -1,6 +1,7 @@
 #pragma once
 #include "Internal.hpp"
 #include "LuaTypes.hpp"
+#include "STDContainers.hpp"
 #include "LuaAux.hpp"
 #include "Property.hpp"
 #include "ClassConstructor.hpp"
